+++ packages/web/src/pages/download-page/DownloadPage.tsx
@@ -10,8 +10,8 @@ import { ParallaxProvider } from 'react-scroll-parallax'
 import IconDownloadDesktop from 'assets/img/publicSite/downloadDesktop.svg'
 import IconDownloadMobile from 'assets/img/publicSite/downloadMobile.svg'
 import { CookieBanner } from 'components/cookie-banner/CookieBanner'
-import Footer from 'components/public-site/FooterV2'
-import NavBanner from 'components/public-site/NavBannerV2'
+import Footer from 'components/public-site/Footer'
+import NavBanner from 'components/public-site/NavBanner'
 import CTAStartListening from 'pages/landing-page/components/CTAStartListening'
 import DownloadApp from 'services/download-app/DownloadApp'
 import { getIOSAppLink } from 'utils/appLinks'
