+++ packages/web/src/pages/terms-of-use-page/TermsOfUsePage.tsx
@@ -1,5 +1,5 @@
-import Footer from 'components/public-site/FooterV2'
-import NavBanner from 'components/public-site/NavBannerV2'
+import Footer from 'components/public-site/Footer'
+import NavBanner from 'components/public-site/NavBanner'
 
 import styles from './TermsOfUsePage.module.css'
